requirements.txt
openai>=1.3.0
python-dotenv>=1.0.0
requests>=2.31.0
anthropic>=0.28.0
pyperclip>=1.8.2
uvicorn>=0.29.0
fastapi>=0.110.0
websockets>=10.0